feat: 整理运行时模块并扩展 MCP 支持
- 将角色模型模块迁移到 avator/，本地 LLM 模块迁移到 llm/，并同步 Gradle、CI 与架构文档路径。

- 引入 transport-neutral MCP runtime session，新增远程 Kotlin runtime、配置导入解析和对应测试。

- 调整 MCP bridge 为本地 stdio 边界，刷新打包后的 bridge assets 与 pnpm workspace/lock。

- 增加 GitHub OAuth 外部浏览器选择与 loopback callback server。

- 统一聊天 token 统计为 Long，并更新消息模型、服务委托与相关测试。

- 补充多项 TODO 计划文档，记录 MCP、GitHub 登录、JS engine 生命周期、模块迁移和 token 统计工作。

+++ app/src/main/java/com/ai/assistance/operit/api/chat/llmprovider/AIService.kt
@@ -10,13 +10,13 @@ import com.ai.assistance.operit.util.stream.Stream
 /** AI服务接口，定义与不同AI提供商进行交互的标准方法 */
 interface AIService {
     /** 输入token计数 (仅新增部分) */
-    val inputTokenCount: Int
+    val inputTokenCount: Long
 
     /** 缓存命中的输入token计数 */
-    val cachedInputTokenCount: Int
+    val cachedInputTokenCount: Long
 
     /** 输出token计数 */
-    val outputTokenCount: Int
+    val outputTokenCount: Long
 
     /** 获取供应商:模型标识符，格式如"DEEPSEEK:deepseek-chat" */
     val providerModel: String
@@ -56,7 +56,7 @@ interface AIService {
             stream: Boolean = true,
             availableTools: List<ToolPrompt>? = null,
             preserveThinkInHistory: Boolean = false, // 新增参数，控制是否保留历史中的思考过程
-            onTokensUpdated: suspend (input: Int, cachedInput: Int, output: Int) -> Unit = { _, _, _ -> },
+            onTokensUpdated: suspend (input: Long, cachedInput: Long, output: Long) -> Unit = { _, _, _ -> },
             onNonFatalError: suspend (error: String) -> Unit = {},
             enableRetry: Boolean = true
     ): Stream<String>
@@ -79,7 +79,7 @@ interface AIService {
     suspend fun calculateInputTokens(
             chatHistory: List<PromptTurn>,
             availableTools: List<ToolPrompt>? = null
-    ): Int
+    ): Long
 
     /**
      * 释放资源
