+++ app/src/main/java/com/ai/assistance/operit/api/chat/llmprovider/RateLimitedAIService.kt
@@ -21,7 +21,7 @@ class RateLimitedAIService(
         stream: Boolean,
         availableTools: List<ToolPrompt>?,
         preserveThinkInHistory: Boolean,
-        onTokensUpdated: suspend (input: Int, cachedInput: Int, output: Int) -> Unit,
+        onTokensUpdated: suspend (input: Long, cachedInput: Long, output: Long) -> Unit,
         onNonFatalError: suspend (error: String) -> Unit,
         enableRetry: Boolean
     ): Stream<String> = com.ai.assistance.operit.util.stream.stream {
